update whitelist num

+++ hardhat-tutorial/contracts/Whitelist.sol
@@ -13,11 +13,11 @@ contract Whitelist {
 
     // numAddressesWhitelisted would be used to keep track of how many addresses have been whitelisted
     // NOTE: Don't change this variable name, as it will be part of verification
-    uint8 public numAddressesWhitelisted;
+    uint256 public numAddressesWhitelisted;
 
     // Setting the Max number of whitelisted addresses
     // User will put the value at the time of deployment
-    constructor(uint8 _maxWhitelistedAddresses) {
+    constructor(uint256 _maxWhitelistedAddresses) {
         maxWhitelistedAddresses =  _maxWhitelistedAddresses;
     }
